tools/syncstorage-loadtests/pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "syncstorage-loadtest"
version = "0.1.0"
description = "Molotov-based load tester for syncstorage-rs"
authors = [
    {name = "Mozilla Sync Backend Engineering", email = "sync-backend@mozilla.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "molotov>=2.6",
    "pexpect>=4.9.0,<5.0.0",
    "configparser>=7.2.0,<8.0.0",
    "hawkauthlib>=2.0.0,<3.0.0",
    "tokenlib>=2.0.0",
    "aiodogstatsd>=0.16.0,<0.17.0",
    "PyFxA>=0.7.0",
    "PyJWT>=2.0.0",
    "cryptography>=40.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.16.0",
    "pydocstyle>=6.3.0",
    "ruff>=0.12.7",
    "black>=25.1.0",
    "bandit>=1.8.3",
    "isort>=6.0.1",
]

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
